refs #2 afegides opcions per a mirar scenes dins d'scenes, recollir objectes i posat de mostra la escena de les pomes i l'encenedor
git-svn-id: http://aproject.unfuddle.com/svn/aproject_androidvj@13 13ae8a00-e42d-4a4a-b426-7c3bbda1a15a

+++ aVideogame/src/avideogame/present/aVideogame.java
@@ -8,6 +8,7 @@ import android.util.Log;
 import android.view.MotionEvent;
 import android.view.Window;
 import avideogame.domain.DomainController;
+import avideogame.utils.Utilities;
 
 
 public class aVideogame extends Activity {
@@ -31,7 +32,7 @@ public class aVideogame extends Activity {
 		Log.d("OnTouchEvent","ZZ");
 		
 		Intent sceneIntent = new Intent(getBaseContext(), SceneActivity.class);
-		sceneIntent.putExtra("SceneIndex", (int)0);
+		sceneIntent.putExtra("SceneIndex", 0);
 		startActivity(sceneIntent);
 		
 		return super.onTouchEvent(event);
